Connects inputted username and password to the MySQL DB
Default values for the following were added:
- Database username
- Database password
- Host name
- Database name/title
- Port
- Socket

There are error checks to see if the connection to the database was successful.
When user enters his/her username and password, it will check with the data
in the Member table in the DB. If there is a matching account, he/she will
be logged into their homepage. If not, user has sees an error message.

TODO: Need to hash (MD5) the password.